RCPP-46 Run tests in SPM, cleanup query & table
