+++ Package.swift
@@ -97,6 +97,25 @@ let package = Package(
             sources: [
                 "main.hpp",
                 "main.cpp",
+                "db/binary_tests.cpp",
+                "db/date_tests.cpp",
+                "db/decimal_tests.cpp",
+                "db/embedded_object_tests.cpp",
+                "db/frozen_tests.cpp",
+                "db/list_tests.cpp",
+                "db/map_tests.cpp",
+                "db/mixed_tests.cpp",
+                "db/numeric_tests.cpp",
+                "db/object_id_tests.cpp",
+                "db/object_tests.cpp",
+                "db/optional_tests.cpp",
+                "db/performance_tests.cpp",
+                "db/query_tests.cpp",
+                "db/realm_tests.cpp",
+                "db/results_tests.cpp",
+                "db/run_loop_tests.cpp",
+                "db/set_tests.cpp",
+                "db/string_tests.cpp"
             ],
             resources: [
                 .copy("setup_baas.rb"),
